Clarify the value of agent_name for AWS runner registration
When using AWS agents we don't have a single agent name, and the
provided example agent_name of "{{build_name}}-{{user `build_number`}}"
doesn't really explain what this value should be.

This was confusing for me when going through the README and trying to
set this up so I assume it will be for others as well.